♻️ DXR7ZX context: decompose init builders
