+++ .agentplane/tasks/202605282234-DXR7ZX/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 DXR7ZX task: Context init builder decomposition [202605282234-DXR7ZX]
