+++ .agentplane/tasks/202605282234-DXR7ZX/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-28T22:34:52.803Z
+
+## Task
+
+- Task: `202605282234-DXR7ZX`
+- Title: Context init builder decomposition
+- Status: DOING
+- Branch: `task/202605282234-DXR7ZX/context-init-builder-decomposition`
+- Canonical task record: `.agentplane/tasks/202605282234-DXR7ZX/README.md`
+
+## Verification
+
+- State: ok
+- Note: Context init was decomposed into command orchestration, bootstrap/git helpers, and content builders. Verified with focused context CLI tests, typecheck, arch deps, lint, format, and hotspot threshold check (runtime warnings 39 -> 38).
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-28T22:34:52.803Z
+- Branch: task/202605282234-DXR7ZX/context-init-builder-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
