+++ packages/agentplane/src/commands/context/init.ts
@@ -1,22 +1,26 @@
 /* eslint-disable @typescript-eslint/no-unused-vars, unicorn/no-array-sort */
-import { execFile } from "node:child_process";
-import { readdir, readFile } from "node:fs/promises";
+import { readFile } from "node:fs/promises";
 import path from "node:path";
-import { promisify } from "node:util";
-
-import { gitAddPaths, gitCommit, gitEnv, gitStagedPaths } from "@agentplaneorg/core/git";
 
 import { infoMessage } from "../../cli/output.js";
-import { exitCodeForError } from "../../cli/exit-codes.js";
-import { cmdInit } from "../../cli/run-cli/commands/init/orchestrate.js";
-import { detectParentGitRoot } from "../../cli/run-cli/commands/init/git.js";
-import type { InitParsed } from "../../cli/run-cli/commands/init/model.js";
-import type { CommandSpec } from "../../cli/spec/spec.js";
 import { CliError } from "../../shared/errors.js";
-import { writeJsonStableIfChanged, writeTextIfChanged } from "../../shared/write-if-changed.js";
-import { loadCommandContext, type CommandContext } from "../shared/task-backend.js";
+import { writeTextIfChanged } from "../../shared/write-if-changed.js";
+import type { CommandContext } from "../shared/task-backend.js";
 
 import type { ContextInitParsed } from "./context.spec.js";
+import {
+  assertContextBootstrapIndexClean,
+  commitContextBootstrapIfChanged,
+  loadOrBootstrapCommandContext,
+} from "./context-init-bootstrap.js";
+import {
+  buildCapabilitiesReadme,
+  buildContextReadme,
+  buildPolicyMarkdown,
+  buildRedactionRulesYaml,
+  buildSyncRulesYaml,
+  buildWikiAgentsMarkdown,
+} from "./context-init-builders.js";
 import {
   assertProfileSwitchIsExplicit,
   POLICY_FILES,
@@ -31,9 +35,6 @@ import { starterWikiPageFiles, wikiFrontmatter } from "./init-wiki.js";
 import { buildWikiPolicyMarkdown } from "./init-wiki-policy.js";
 import { cmdContextReindex } from "./reindex.js";
 
-const execFileAsync = promisify(execFile);
-const CONTEXT_BOOTSTRAP_TASK_ID = "202601010101-CTX1NT";
-
 const DEFAULT_GITIGNORE_ENTRIES = [
   ".agentplane/cache.sqlite",
   ".agentplane/cache.sqlite-wal",
@@ -43,15 +44,6 @@ const DEFAULT_GITIGNORE_ENTRIES = [
   ".agentplane/context/service/embeddings/",
   ".agentplane/context/service/remotes/",
 ];
-
-const SAFE_EMPTY_PLACEHOLDERS = new Set([".DS_Store"]);
-
-const contextBootstrapInitSpec: CommandSpec<InitParsed> = {
-  id: ["init"],
-  group: "Setup",
-  summary: "Initialize agentplane project files before context bootstrap.",
-  parse: () => ({ yes: true }),
-};
 
 type InitReport = {
   created: string[];
@@ -115,177 +107,6 @@ export async function cmdContextInit(opts: {
       message: `context init failed: ${(err as Error).message}`,
     });
   }
-}
-
-async function loadOrBootstrapCommandContext(opts: {
-  cwd: string;
-  rootOverride?: string | null;
-}): Promise<{ ctx: CommandContext; bootstrapped: boolean }> {
-  const root = path.resolve(opts.rootOverride ?? opts.cwd);
-  const target = await inspectBootstrapTarget(root);
-  if (target?.canBootstrap) {
-    return { ctx: await bootstrapEmptyProjectForContextInit(root), bootstrapped: true };
-  }
-
-  try {
-    return {
-      ctx: await loadCommandContext({ cwd: opts.cwd, rootOverride: opts.rootOverride ?? null }),
-      bootstrapped: false,
-    };
-  } catch (err) {
-    if (target && !target.hasAgentplaneDir && isMissingProjectError(err)) {
-      throw new CliError({
-        exitCode: exitCodeForError("E_USAGE"),
-        code: "E_USAGE",
-        message:
-          "context init can bootstrap AgentPlane only in an empty directory. " +
-          "Run agentplane init explicitly before context init for non-empty directories.",
-      });
-    }
-    throw err;
-  }
-}
-
-async function assertContextBootstrapIndexClean(root: string): Promise<void> {
-  const baseGitEnv = gitEnv();
-  if (!(await isInsideGitWorkTree(root, baseGitEnv))) return;
-  const stagedBefore = await gitStagedPaths(root);
-  if (stagedBefore.length === 0) return;
-  throw new CliError({
-    exitCode: exitCodeForError("E_GIT"),
-    code: "E_GIT",
-    message:
-      "Git index has staged changes; commit or unstage them before running agentplane context init.",
-  });
-}
-
-async function commitContextBootstrapIfChanged(root: string, paths: string[]): Promise<boolean> {
-  const baseGitEnv = gitEnv();
-  if (!(await isInsideGitWorkTree(root, baseGitEnv))) return false;
-  const dedupedPaths = [...new Set(paths)].filter((entry) => entry.length > 0);
-  const committablePaths = await filterGitIgnoredPaths(root, dedupedPaths, baseGitEnv);
-  if (committablePaths.length === 0) return false;
-  await gitAddPaths(root, committablePaths);
-  const staged = await gitStagedPaths(root);
-  if (staged.length === 0) return false;
-  const env = {
-    ...baseGitEnv,
-    AGENTPLANE_ALLOW_POLICY: "1",
-    AGENTPLANE_TASK_ID: CONTEXT_BOOTSTRAP_TASK_ID,
-  };
-  await gitCommit(
-    root,
-    [
-      "✅ CTX1NT task: initialize AgentPlane context",
-      "",
-      "Context-Bootstrap: true",
-      `Context-Bootstrap-Task: ${CONTEXT_BOOTSTRAP_TASK_ID}`,
-    ].join("\n"),
-    { env },
-  );
-  return true;
-}
-
-async function isInsideGitWorkTree(root: string, env: NodeJS.ProcessEnv): Promise<boolean> {
-  try {
-    const result = await execFileAsync("git", ["rev-parse", "--is-inside-work-tree"], {
-      cwd: root,
-      env,
-    });
-    return result.stdout.trim() === "true";
-  } catch {
-    return false;
-  }
-}
-
-async function filterGitIgnoredPaths(
-  root: string,
-  paths: string[],
-  env: NodeJS.ProcessEnv,
-): Promise<string[]> {
-  const kept: string[] = [];
-  for (const relative of paths) {
-    if (await isGitIgnoredPath(root, relative, env)) continue;
-    kept.push(relative);
-  }
-  return kept;
-}
-
-async function isGitIgnoredPath(
-  root: string,
-  relative: string,
-  env: NodeJS.ProcessEnv,
-): Promise<boolean> {
-  try {
-    await execFileAsync("git", ["check-ignore", "--quiet", "--", relative], { cwd: root, env });
-    return true;
-  } catch (err) {
-    const code = (err as { code?: number | string } | null)?.code;
-    if (code === 1) return false;
-    throw err;
-  }
-}
-
-type BootstrapTarget = {
-  canBootstrap: boolean;
-  hasAgentplaneDir: boolean;
-};
-
-async function inspectBootstrapTarget(root: string): Promise<BootstrapTarget | null> {
-  const entries = await readBootstrapTargetEntries(root);
-  if (!entries) return null;
-  const meaningfulEntries = entries.filter((entry) => !SAFE_EMPTY_PLACEHOLDERS.has(entry));
-  const canBootstrap =
-    meaningfulEntries.length === 0 ||
-    (meaningfulEntries.length === 1 && meaningfulEntries[0] === ".git");
-
-  return {
-    canBootstrap,
-    hasAgentplaneDir: meaningfulEntries.includes(".agentplane"),
-  };
-}
-
-async function bootstrapEmptyProjectForContextInit(root: string): Promise<CommandContext> {
-  const parentGitRoot = await detectParentGitRoot(root);
-  if (parentGitRoot) {
-    throw new CliError({
-      exitCode: exitCodeForError("E_USAGE"),
-      code: "E_USAGE",
-      message:
-        "context init can bootstrap AgentPlane only in a standalone empty directory. " +
-        `Target ${root} is inside parent Git repository ${parentGitRoot}. ` +
-        "Run agentplane init explicitly from the intended nested project root, then run agentplane context init.",
-    });
-  }
-
-  process.stdout.write(
-    infoMessage("agentplane project not found; bootstrapping empty directory") + "\n",
-  );
-  await cmdInit({
-    cwd: root,
-    rootOverride: root,
-    outputMode: "text",
-    flags: { yes: true },
-    spec: contextBootstrapInitSpec,
-  });
-  return await loadCommandContext({ cwd: root, rootOverride: root });
-}
-
-async function readBootstrapTargetEntries(root: string): Promise<string[] | null> {
-  try {
-    return await readdir(root);
-  } catch {
-    return null;
-  }
-}
-
-function isMissingProjectError(err: unknown): boolean {
-  const message = err instanceof Error ? err.message : String(err);
-  return (
-    message.startsWith("Not a git repository") ||
-    message.includes("ENOENT") ||
-    message.includes(".agentplane")
-  );
 }
 
 async function createContextWorkspace(
@@ -428,135 +249,6 @@ async function ensureContextGitignore(root: string, parsed: ContextInitParsed): 
   if (currentText.trimEnd() === nextText.trimEnd()) return false;
   await writeTextIfChanged(gitignorePath, nextText);
   return true;
-}
-
-function buildContextReadme(profile: ContextInitParsed["profile"]): string {
-  const maximumAssimilation =
-    profile === "maximum-assimilation"
-      ? `
-Maximum-assimilation mode adds a stricter wiki maintenance contract:
-
-- Preserve all significant source meaning in wiki, facts, graph, glossary, and coverage
-  artifacts so maintained context remains useful even if \`context/raw/**\` is later removed.
-- Keep original source identity in a source registry with \`source_id\`, original path,
-  \`sha256:\` hash, content type, line count, ingest time, and availability state.
-- Use line-addressed source refs as provenance pointers, not retained content; missing raw refs may be non-dereferenceable while wiki/fact/graph artifacts stay self-contained.
-- Extract entities, aliases, relations, decisions, requirements, risks, workflows, and conflicts
-  before writing narrative articles.
-- Choose the wiki structure from source content; do not create the default
-  concepts/entities/decisions/modules/contradictions/reports scaffold unless source analysis
-  justifies it. Record the topology decision before creating page families.
-- Create or maintain the canonical glossary as root file \`context/wiki/glossary.md\`; keep it as navigation/aliases over wiki pages and graph entities, then use canonical terms in prose.
-- Create Obsidian page properties automatically for wiki pages: \`aliases\`, \`tags\`, and
-  \`cssclasses\` stay alongside AgentPlane frontmatter for better vault rendering.
-- Use Obsidian-compatible links whose target case exactly matches a canonical page path, title, or
-  alias; prefer \`[[canonical-page|Display Label]]\` when file path and display title differ.
-- Cite raw sources in prose with numeric notes like \`[1]\`; keep the raw-data Markdown links in a
-  trailing \`## Sources\` section.
-- Treat coverage gaps, unresolved entity identity, sensitive-source leakage risk, and missing line refs
-  as blockers or explicit approval-required findings.
-- Require EVALUATOR review of topology, granularity, wikilinks, coverage, glossary safety,
-  raw-deletion resilience, and private leakage risk before finish.
-`
-      : "";
-  return `# Context workspace
-
-Profile: ${profile}
-
-Use this directory as the human-readable context surface.
-
-AgentPlane local context is optional and independent from runner prompt assembly. When enabled, it
-uses one llm-wiki contract:
-
-- \`context/raw/**\` keeps source material.
-- \`context/wiki/**\` keeps readable synthesis pages with AgentPlane frontmatter.
-- \`.agentplane/context/derived/**\` keeps reproducible claims, graph rows, provenance, and reports.
-- \`.agentplane/context/service/**\` keeps local caches only.
-
-Agents should create wiki pages when a topic is reusable for future tasks, but keep atomic claims in
-derived machine artifacts. \`context init\` creates only \`context/raw/.gitkeep\`,
-\`context/wiki/AGENTS.md\`, and \`context/wiki/index.md\`; users own any hierarchy below
-\`context/raw/**\`. Source references should preserve user-created raw paths where possible.
-${maximumAssimilation}
-`;
-}
-
-function buildWikiAgentsMarkdown(profile: ContextInitParsed["profile"]): string {
-  const maximumAssimilation =
-    profile === "maximum-assimilation"
-      ? `
-## Maximum assimilation mode
-
-- Use the \`context.maximum_assimilation\` blueprint for new context assimilation tasks.
-- Goal: after assimilation, the maintained wiki and derived artifacts preserve all significant
-  source meaning without relying on raw files for semantic recall.
-- Keep original hashes in the source-set lock and cite source content with concrete line refs such as
-  \`context/raw/<user-path>/note.md#lines=12-24\`; treat those refs as audit provenance, not as the
-  stored meaning.
-- First pass: build or update canonical entities, glossary aliases, relation candidates, conflicts, and coverage notes.
-- Glossary output: create or update \`context/wiki/glossary.md\` as the root glossary file; do not scatter canonical glossary state across reports or page-local notes.
-- Topology pass: choose wiki structure from source content; do not mechanically create
-  \`concepts/\`, \`entities/\`, \`decisions/\`, \`modules/\`, \`contradictions/\`, or \`reports/\`.
-- Record a topology decision before page-family creation. It must classify the source shape (book/corpus, codebase, task history, product docs, research notes, ops logs, or another named shape), name canonical page families, justify page-vs-heading granularity, map source-local terms to canonical labels or aliases, and keep ambiguous identities as open questions.
-- Second pass: synthesize granular wiki articles from that graph/glossary layer; use canonical terms from \`context/wiki/glossary.md\` and preserve source-local wording as aliases.
-- Create separate pages for reusable entities, concepts, decisions, requirements, risks, workflows,
-  and modules; use stable headings for smaller objects inside broader pages.
-- Ensure each wiki page has Obsidian properties \`aliases\`, \`tags\`, and \`cssclasses\` in YAML
-  frontmatter in addition to the \`agentplane_context\` manifest.
-- Use Obsidian-compatible wikilinks whose target case exactly matches a canonical page path, title,
-  or alias; add a display alias or heading anchor only when display wording differs.
-- Cite raw sources in prose with numeric notes like \`[1]\`; keep Markdown raw-data links in a
-  trailing \`## Sources\` section rather than scattering long source URLs through prose.
-- Record extraction coverage: covered source spans, intentionally omitted boilerplate, redacted
-  spans, unresolved conflicts, and open questions.
-- Record EVALUATOR review for topology, granularity, wikilinks, line refs, glossary safety,
-  coverage gaps, raw-deletion resilience, and private leakage risk before finish.
-- If a raw source is missing later, keep its source registry entry with availability state
-  \`missing\`; the wiki, facts, graph, glossary, and coverage artifacts must still carry the
-  assimilated meaning.
-- Do not copy secrets into public wiki pages. Redact sensitive source spans before publication.
-`
-      : "";
-  return `${wikiFrontmatter("wiki.agents", "Context wiki agent notes", "policy")}
-
-# Context wiki agent notes
-
-Profile: ${profile}
-
-- Treat \`context/wiki/**\` as durable, source-backed project knowledge with stable AgentPlane frontmatter.
-- Treat \`.agentplane/context/agentplane.context.yaml\` as the machine-readable context contract and \`.agentplane/context/policies/wiki.rules.md\` as the human-readable wiki policy.
-- Analyze the base project, existing docs, task history, and raw sources before choosing a wiki structure.
-- Choose the smallest wiki hierarchy that fits this project; do not force a universal concepts/entities/decisions/modules layout.
-- Keep this initialized wiki minimal until first ingest; project-specific folders should appear from source-backed assimilation, not from empty scaffolding.
-- Preserve modality, source refs, cross-links, glossary aliases, and graph alignment when updating pages.
-- Write synthesized wiki prose in English by default; preserve source-language terms only for quotes, titles, proper names, aliases, paths, and code identifiers.
-- Prefer updating existing canonical pages over creating duplicates; describe small objects under stable headings when that is clearer.
-- Use \`agentplane context wiki new\`, \`agentplane context wiki lint\`, \`agentplane context wiki explain\`, \`agentplane context wiki link\`, and \`agentplane context wiki index\`.
-- When claims conflict, keep both claims, create a conflict candidate, and ask for review before promotion or overwrite.
-- Keep raw inputs in \`context/raw/**\`; preserve the user-created hierarchy when citing sources.
-- Add source references for factual claims and run \`agentplane context verify-task <task-id>\` before closing context assimilation work.
-${maximumAssimilation}
-
-## Sources
-
-- no-source: generated policy notes from \`agentplane context init\`; add source references before promotion.
-`;
-}
-
-function buildCapabilitiesReadme(profile: ContextInitParsed["profile"]): string {
-  return `# Context capabilities\n\nProfile: ${profile}\n\nReusable artefacts for prompts, playbooks, templates, checklists and rubrics.\n`;
-}
-
-function buildPolicyMarkdown(name: string): string {
-  return `# ${name}\n\n- Keep raw sources in \`context/raw\`.\n- Keep durable machine artifacts under \`.agentplane/context/derived\`.\n- Keep service caches under \`.agentplane/context/service\`.\n`;
-}
-
-function buildRedactionRulesYaml(): string {
-  return `mode: explicit\nallowlist: []\ndenylist: []\n`;
-}
-
-function buildSyncRulesYaml(): string {
-  return `mode: manual\nallow_external: false\n`;
 }
 
 async function readExisting(filePath: string): Promise<string | null> {
